# Manager: Optimize getPanels function with memoization